Add standalone remote_sync package
